Moving types into separate file to remove circular dependency
