+++ src/Types.tsx
@@ -1,0 +1,19 @@
+export enum Mode {
+  Normal = 'NORMAL',
+  Assay = 'ASSAY'
+}
+
+export enum CellPart {
+  Nucleus = 'NUCLEUS',
+  Cytoplasm = 'CYTOPLASM',
+  Golgi = 'GOLGI',
+  Gates = 'GATES',
+  Intercell = 'INTERCELL',
+  None = 'NONE'
+}
+
+export enum Substance {
+  Substance1 = 'Substance 1',
+  Substance2 = 'Substance 2',
+  Substance3 = 'Substance 3'
+}
